requirements.txt
Django==4.1.2
django-ckeditor==6.5.1
Pillow==9.3.0
